version 11.3.3

+++ common_java/src/main/java/com/testerkit/common/json/PointJson.java
@@ -18,7 +18,7 @@ public class PointJson {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -26,7 +26,7 @@ public class PointJson {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
